20220614/push_sub/requirements.txt
flask
google-cloud-firestore
google-cloud-pubsub
gunicorn
